fix: crash when setting the diff code
